+++ vllm/worker/spec_decode/interfaces.py
@@ -16,13 +16,45 @@ from vllm.model_executor.layers.rejection_sampler import RejectionSampler
 from vllm.model_executor.parallel_utils.parallel_state import get_tensor_model_parallel_group
 from vllm.config import CacheConfig
 from vllm.utils import in_wsl
-from vllm.worker.spec_decode.util import nvtx_range, sampler_output_to_torch, SpeculativeProposals, get_all_seq_ids
+from vllm.worker.spec_decode.util import nvtx_range, sampler_output_to_torch, get_all_seq_ids
 
 
 logger = logging.getLogger(__name__)
 
 from abc import ABC, abstractmethod
 
+
+@dataclass
+class SpeculativeProposals:
+    """Sequences valid for speculative decoding and their corresponding
+    speculative proposals generated by the draft worker.
+    """
+    # Proposal token ids and probs speculated by draft worker.
+    proposal_token_ids: torch.Tensor
+    proposal_probs: torch.Tensor
+    proposal_lens: torch.Tensor
+
+    def __repr__(self):
+        return (f"SpeculativeProposals("
+                f"proposal_token_ids={self.proposal_token_ids.shape}, "
+                f"proposal_probs={self.proposal_probs.shape}, "
+                f"proposal_lens={self.proposal_lens.shape})")
+
+@dataclass
+class SpeculativeScores:
+    """Sequences valid for speculative decoding and their corresponding
+    speculative proposals generated by the draft worker.
+    """
+    # Proposal token ids and probs speculated by draft worker.
+    proposal_token_ids: torch.Tensor
+    proposal_probs: torch.Tensor
+    proposal_lens: torch.Tensor
+
+    def __repr__(self):
+        return (f"SpeculativeProposals("
+                f"proposal_token_ids={self.proposal_token_ids.shape}, "
+                f"proposal_probs={self.proposal_probs.shape}, "
+                f"proposal_lens={self.proposal_lens.shape})")
 
 class SpeculativeProposer(ABC):
     @abstractmethod
