+++ vllm/worker/spec_decode/multi_step_worker.py
@@ -5,7 +5,8 @@ import torch
 
 from vllm.sequence import SamplerOutput, SequenceGroupMetadata
 from vllm.worker.worker import Worker
-from vllm.worker.spec_decode.util import SpeculativeProposals, sampler_output_to_torch
+from vllm.worker.spec_decode.util import sampler_output_to_torch
+from vllm.worker.spec_decode.interfaces import SpeculativeProposals
 from vllm.worker.spec_decode.proposing import DraftModelTop1Proposer
