+++ vllm/worker/spec_decode/util.py
@@ -7,22 +7,6 @@ from itertools import chain
 
 SeqId = int
 
-@dataclass
-class SpeculativeProposals:
-    """Sequences valid for speculative decoding and their corresponding
-    speculative proposals generated by the draft worker.
-    """
-    # Proposal token ids and probs speculated by draft worker.
-    proposal_token_ids: torch.Tensor
-    proposal_probs: torch.Tensor
-    proposal_lens: torch.Tensor
-
-    def __repr__(self):
-        return (f"SpeculativeProposals("
-                f"proposal_token_ids={self.proposal_token_ids.shape}, "
-                f"proposal_probs={self.proposal_probs.shape}, "
-                f"proposal_lens={self.proposal_lens.shape})")
-
 
 def get_all_seq_ids(seq_group_metadata_list: List[SequenceGroupMetadata]) -> List[SeqId]:
     """Given a list of SequenceGroupMetadata, create a list of all
